+++ src/components/pages/apps/network/svg/network-svg.tsx
@@ -36,8 +36,8 @@ export function NetworkSvgContent() {
     groups,
     coordinates,
     size: d3Size,
-    sizeLim,
-    sizeLim2,
+    metricLim1,
+    metricLim2,
   } = useNetwork()
 
   // const { showTooltip, hideTooltip } = useTooltip()
@@ -99,14 +99,14 @@ export function NetworkSvgContent() {
     const radiusMap = new Map<string, number>(
       network.nodes.map((node) => [
         node.id,
-        ((node.size ?? 0) / sizeLim.max) * settings.plot.nodes.radius,
+        ((node.size ?? 0) / metricLim1.max) * settings.plot.nodes.radius,
       ])
     )
 
     const sizeMap2 = new Map<string, number>(
       network.nodes.map((node) => [
         node.id,
-        (node.size2 ?? 0) / (sizeLim2?.max ?? 1),
+        (node.size2 ?? 0) / (metricLim2?.max ?? 1),
       ])
     )